address comments in Eric's review
