--- tiledb/common/path.h
@@ -1,73 +1,0 @@
-/**
- * @file   path.h
- *
- * @section LICENSE
- *
- * The MIT License
- *
- * @copyright Copyright (c) 2017-2022 TileDB, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @section DESCRIPTION
- *
- * This file implements class path.
- */
-
-#ifndef TILEDB_PATH_H
-#define TILEDB_PATH_H
-
-#include <string>
-
-namespace tiledb {
-namespace common {
-namespace filesystem {
-
-class path {
- public:
-  /* ********************************* */
-  /*     CONSTRUCTORS & DESTRUCTORS    */
-  /* ********************************* */
-
-  /** Constructor. */
-  path() = default;
-
-  /**
-   * Constructor
-   *
-   * @param p The path to hold
-   */
-  path(const std::string& p);
-
-  /* ********************************* */
-  /*                API                */
-  /* ********************************* */
-
-  /** Returns the internal path string*/
-  const std::string& native() const;
-
- private:
-  std::string path_;
-};
-
-}  // namespace filesystem
-}  // namespace common
-}  // namespace tiledb
-
-#endif  // TILEDB_PATH_H
