+++ tiledb/sm/filesystem/win.cc
@@ -33,7 +33,7 @@
 
 #include "win.h"
 #include "path_win.h"
-#include "tiledb/common/directory_entry.h"
+#include "tiledb/common/filesystem/directory_entry.h"
 #include "tiledb/common/heap_memory.h"
 #include "tiledb/common/logger.h"
 #include "tiledb/common/stdx_string.h"
